requirements.txt
openai
rich>=13.7.0
click~=8.1.7
inquirer~=3.2.4
pyperclip~=1.8.2
chromadb~=0.4.24
psutil>=5.9.0
instructor~=0.6.8
pydantic>=2.6.1
setuptools~=68.2.2
python-nmap>=0.7.1
requests>=2.31.0
python-ollama>=0.1.6
typer>=0.9.0
python-dotenv>=1.0.0
